requirements.txt
# RLCHF Requirements
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
accelerate>=0.20.0
peft>=0.4.0
trl>=0.7.0

# Inference and model serving
vllm>=0.3.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Configuration and environment
python-dotenv>=1.0.0

# Evaluation and metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0

# Development tools (optional)
black>=23.0.0
pytest>=7.0.0

# Optional dependencies for specific features
# jupyter>=1.0.0  # For notebooks
# matplotlib>=3.7.0  # For plotting
# seaborn>=0.12.0  # For visualization
